[DAG] Implement or version of (and (setne X, 0), (setne X, -1))
The or version, as per DeMorgan's Law, is the inverse of both sides.

This is (or (seteq X, 0), (seteq X, -1)) --> (setult (add X, 1), 2)

Proof:
The inverse of (a and b) is not a or not b, per DeMorgan's Law.

not setne(X, 0) is seteq(X, 0) and same for -1. For the other side, ge's opposite is lt.