Disable outlining for bundles
Currently the vectors do not take bundles into account, causing it to outline bundled instructions to different functions in some cases. Until a proper fix is found that ensures that the outlined function candidate is only legal if all bundled instructions will be outlined if any of them are outlined, which is information the conversion to an unsigned vec removes, it is best to consider all bundled instructions unsafe to outline.